refactor: 재사용하는 컴포넌트 경로 변경
